requirement.txt
# Core LLM and RAG dependencies
langchain>=0.1.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
tiktoken>=0.5.0
transformers>=4.30.0
torch>=2.0.0

# Document processing
unstructured>=0.10.0
unstructured[local-inference]>=0.10.0
PyMuPDF>=1.23.0
python-docx>=0.8.11

# Multilingual support
googletrans>=4.0.0
langdetect>=1.0.9
indic-transliteration>=2.3.0

# Speech processing
speechrecognition>=3.10.0
pyttsx3>=2.90
pydub>=0.25.1

# Web interface
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Database and storage
sqlite3
sqlalchemy>=2.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
google-cloud-sql-connector>=1.4.0

# Google Cloud dependencies
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# Utilities
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Screen reading accessibility
pyautogui>=0.9.54
pyperclip>=1.8.2

